create another script for the qiskit_bot
